+++ lnurl/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Revision history for lnurl
 
+## 0.1.0.1 -- 2022-07-29
+
+* Updates dependencies
+* Changes the `metadatada` field in the payment flow success response to text so
+  that we can calculate hashes over it correctly
+
 ## 0.1.0.0 -- 2021-07-29 
 
 * First version. Released on an unsuspecting world.